// src/api/models/UpdatePunishmentDictDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdatePunishmentDictDto = {
    /**
     * 显示标签
     */
    label?: string;
    /**
     * 描述说明
     */
    description?: string;
    /**
     * 是否启用
     */
    enabled?: boolean;
    /**
     * 排序权重
     */
    sort?: number;
};